esp_hal_lcd: fix float to double promotion warning
The dpi divider trace passes float arguments to a variadic log
macro. The disabled-log path still type-checks the call, so the
implicit promotion warns under -Wdouble-promotion. Cast at the
call site, matching the phy pll trace above it.

+++ components/esp_hal_lcd/mipi_dsi_hal.c
@@ -271,6 +271,7 @@ uint32_t mipi_dsi_hal_host_dpi_calculate_divider(mipi_dsi_hal_context_t *hal, fl
     hal->expect_dpi_clock_freq_mhz = expect_dpi_clk_mhz;
     hal->real_dpi_clock_freq_mhz = clk_src_mhz / (float)div;
     HAL_LOGD(TAG, "dpi clk: src=%.2f MHz, expect=%.2f MHz, div=%" PRIu32 ", real=%.2f MHz",
-             clk_src_mhz, expect_dpi_clk_mhz, div, hal->real_dpi_clock_freq_mhz);
+             (double)clk_src_mhz, (double)expect_dpi_clk_mhz, div,
+             (double)hal->real_dpi_clock_freq_mhz);
     return div;
 }
